Move configs from `root` to `src`
